+++ src/manifolds/HyperbolicHyperboloid.jl
@@ -6,7 +6,7 @@ We only have to correct for the metric, which means that the sign of the last en
 for the result ``Y``  we are looking for a tangent vector such that
 
 ```math
-    g_p(Y,Z) = -y_{n+1}z_{n+1} + \sum_{i=1}^n y_iz_i = \sum_{i=1}^{n+1} z_ix_i
+    g_p(Y,Z) = -y_{n+1}z_{n+1} + \sum_{i=1}^n y_i z_i = \sum_{i=1}^{n+1} z_ix_i
 ```
 
 holds, which directly yields ``y_i=x_i`` for ``i=1,\ldots,n`` and ``y_{n+1}=-x_{n+1}``.