+++ src/manifolds/SymplecticGrassmannProjector.jl
@@ -14,7 +14,7 @@ function check_point(M::SymplecticGrassmann, p::ProjectorPoint; kwargs...)
     if !isapprox(c, p.value; kwargs...)
         return DomainError(
             norm(c - p.value),
-            "The poin $(p) is not equal to its square $c, so it does not lie on $M.",
+            "The point $(p) is not equal to its square $c, so it does not lie on $M.",
         )
     end
     if !isapprox(p.value, symplectic_inverse(p.value); kwargs...)
@@ -56,7 +56,7 @@ function check_vector(
         return DomainError(
             norm((Hamiltonian(X.value)^+).value - X.value),
             (
-                "The matrix X is not in the tangent space at $p of $M, since X is not equal to ist symplectic inverse."
+                "The matrix X is not in the tangent space at $p of $M, since X is not equal to its symplectic inverse."
             ),
         )
     end